serial_station: Batch 61 — 启用 RUF 系列 ruff-specific + 12 处修复
- pyproject.toml [tool.ruff.lint].select 加 RUF
- RUF005/007/015/059 等 12 处修复（自动 + unsafe-fixes）
- waveform_engine.py OpenGL import → importlib.util.find_spec
- ignore RUF001/002/003（中文 Unicode 警告 7900+）、RUF012（mutable class var）
- 恢复 dashboard_panel QFileDialog noqa（RUF unsafe-fix 误删）

lint 现在 F+UP+B+SIM+RUF 五系列强制。全量 pytest 1699 passed，lint clean，smoke exit 0。

评分: 734 + 2 = 736
变更: 16 files, 30 insertions, 30 deletions

+++ python/embeddebug/serial_station/ui/controls/progress_ring.py
@@ -228,7 +228,7 @@ class ProgressRing(QWidget):
             )
             ratio = (disp_clamped - float(self._minimum)) / float(span)
             ratio = max(0.0, min(1.0, ratio))
-            percent = int(round(ratio * 100))
+            percent = round(ratio * 100)
             text = f"{percent}%"
             font = QFont()
             font.setPointSizeF(max(7.0, side / 6.0))
